Add window "stick" and "unstick" commands
Marking a window as sticky keeps it in its current frame so it will
never be cycled into another frame via next, prev, or when creating
a new split.  A sticky window can be moved around in the frame
stack, but will always remain in that frame until it is explicitly
moved to another frame via exchange* commands.